pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "feedback-prompt-learning"
version = "0.1.0"
description = "An evaluation-driven framework for automatic prompt optimization with rich textual feedback"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
    {name = "Chingis Oinar", email = "your.email@example.com"},
]
keywords = ["prompt-optimization", "llm", "evaluation", "feedback-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "langchain-core",
    "langchain-openai",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "pytest>=9.0.1",
    "torch>=2.9.1",
    "nltk>=3.9.2",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.4.0",
]

[project.urls]
Homepage = "https://github.com/cowana-ai/feedback-prompt-learning"
Repository = "https://github.com/cowana-ai/feedback-prompt-learning"
Issues = "https://github.com/cowana-ai/feedback-prompt-learning/issues"

[tool.hatch.build.targets.wheel]
packages = ["feedback_prompt_learning"]
